requirements.txt
mysql-connector==2.2.9
numpy==1.18.4
pandas==1.0.3
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0